% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_BIC.R
\name{AIC_BIC}
\alias{AIC_BIC}
\title{Calculate AIC and BIC for an lm Model}
\usage{
AIC_BIC(model)
}
\arguments{
\item{model}{A fitted model from ols_fit}
}
\value{
A list containing:
\describe{
  \item{AIC}{The Akaike Information Criterion.}
  \item{BIC}{The Bayesian Information Criterion.}
}
}
\description{
This function computes the Akaike Information Criterion (AIC) and
Bayesian Information Criterion (BIC) for a fitted linear model created using `lm`.
}
