% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_of_variance.R
\name{analysis_of_variance}
\alias{analysis_of_variance}
\title{Analysis of Variance Function}
\usage{
analysis_of_variance(data, response, covariate)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{response}{The response variable (numeric).}

\item{covariate}{The grouping variable (factor or character).}
}
\value{
A list with ANOVA components: TSS, WSS, BSS, F-statistic, p-value, and a summary table.
}
\description{
This function calculates an ANOVA table and p-value.
}
\examples{
analysis_of_variance(iris, Sepal.Length, Species)
}
