% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA}
\usage{
ANOVA(formula, data, type)
}
\arguments{
\item{formula}{An object of class "formula": a symbolic description of the model to be fitted. A typical model has the form outcome ~ covariates
where outcome is the numeric response vector (which people usually denote as Y in statistical formula) and covariates are predictor of response.}

\item{data}{A data frame (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{type}{the type of sums of squares to be obtained. Use "Sequential" for sequential sums of squares and "Partial" for partial sums of squares.}
}
\value{
ANOVA returns ANOVA table in data.frame.
}
\description{
ANOVA is used to obtain ANOVA table of a fitted model. It yields the same results as anova(lm( )) for sequential sums of squares
and car::Anova(lm( ), type="III") for partial sums of squares.
}
\examples{

ANOVA(mpg ~ cyl + wt, mtcars, type = "Partial") ## Get partial SS
ANOVA(mpg ~ cyl + wt + disp, mtcars, type = "Sequential") ## Get sequential SS

}
