% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_ols_cat}
\alias{anova_ols_cat}
\title{Analysis of Variance Function for Categorical covariate.}
\usage{
anova_ols_cat(data, response, covariate)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{response}{The response variable (numeric).}

\item{covariate}{The grouping variable must be categorical, p=1 only.}
}
\value{
A list with ANOVA components: TSS, WSS, BSS, F-statistic, p-value, and a summary table.
}
\description{
This function calculates a One-Way ANOVA table.
}
