% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coef method of linreg class}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{Linreg class object}

\item{...}{additional parameter}
}
\value{
Returns the coefficients as a named vector.
}
\description{
Coef method of linreg class
}
