% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R, R/RidgeRegression.R
\name{coef}
\alias{coef}
\alias{coef.linreg}
\alias{coef.ridgereg}
\title{Coefficients method for linreg class}
\usage{
coef(x)

\method{coef}{linreg}(x)

coef(x)

\method{coef}{ridgereg}(x)
}
\arguments{
\item{x}{An S3 object of class ridgereg.}
}
\value{
A named vector of coefficients.

A named vector of coefficients.
}
\description{
Coefficients method for linreg class

Coefficients method for ridgereg class
}
