% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coefficients method for ridgereg class}
\usage{
\method{coef}{ridgereg}(x)
}
\arguments{
\item{x}{An S3 object of class ridgereg.}
}
\value{
A named vector of coefficients.
}
\description{
Coefficients method for ridgereg class
}
