% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{confints_tests_coefficients}
\alias{confints_tests_coefficients}
\title{Confidence Intervals and Tests for Individual Coefficients.}
\usage{
confints_tests_coefficients(model, alpha = 0.05, b = 0, bonf = F)
}
\arguments{
\item{model}{Returned object from ols_fit}

\item{alpha}{Significance level}

\item{b}{vector or scalar representing the value(s) to use for the null hypotheses.
(If vector, it's expected to have the same length as the beta vector.)
   That is, the null hypothesis for the betaj hypothesis test is
   H0: betaj = b (if scalar) or H0: betaj = b[j] (if vector).}

\item{bonf}{whether to do Bonferroni correction for multiple comparisons}
}
\value{
A named matrix with the left bound, right bound, p-value for test,
 and adjusted p-value
}
\description{
Confidence Intervals and Tests for Individual Coefficients.
}
