% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{linearReg}
\alias{linearReg}
\title{Linear Regression Model}
\usage{
linearReg(y, X)
}
\arguments{
\item{y}{\itemize{
\item numeric response variable vector (dim = 1 x p)
}}

\item{X}{\itemize{
\item matrix of explanatory variables included in the model (dim= n x p)
}}
}
\value{
Linear regression model summary including residual quantiles,
regression coefficients, standard error, t-statistics, p-values,
95\% confidence intervals, R-squared and Adjusted R-squared, and
varaince inflation factors.
}
\description{
This function loads in a matrix of explanatory variables and a response
vector. The matrix of explanatory variables in converted into a design matrix
and the response vector is converted into a column vector.
}
\examples{
linearReg(as.vector(mtcars["mpg"]),as.matrix(mtcars[c("cyl", "wt")]))

}
