\name{linearRegression-package}
\alias{linearRegression-package}
\alias{linearRegression}
\docType{package}
\title{
\packageTitle{linearRegression}
}
\description{
\packageDescription{linearRegression}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linearRegression}
\packageIndices{linearRegression}
The linreg class provided by this package runs a linear regression by OLS on a given formula and dataset. Important fields such as residuals, predicted values and estimated coefficients can be accessed by calling their respective methods. Additionally, a summary of the model can be printed along with some basic plots. 
}
\author{
\packageAuthor{linearRegression}

Maintainer: \packageMaintainer{linearRegression}
}
\keyword{package, OLS}