% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg}
\alias{linreg-class}
\title{Class of linear regression models using QR decompoisiton method}
\description{
Class of linear regression models using QR decompoisiton method
}
\section{Attributes}{

\itemize{
     \item{\code{coefficients}: vector of coefficients}
     \item{\code{predicted}: vector of predicted values}
     \item{\code{residuals}: vector of residuals}
   }
}

\section{Methods}{

\itemize{
     \item{\code{coef}: returns the vector of residuals \code{e}.}
     \item{\code{resid}: returns the predicted values \code{y}.}
     \item{\code{pred}: return the coefficients as a named vector.}
   }
}
\examples{
#linreg$new(data = iris, formula = iris$Sepal.Length ~ iris$Sepal.Width)
#linreg$new(data = faithful, formula = faithful$eruption ~ faithful$waiting)
}
\references{
QR Decomposition - \url{https://en.wikipedia.org/wiki/QR_decomposition}
}

