\name{linreg-package}
\alias{linreg-package}
\alias{linreg}
\docType{package}
\title{
\packageTitle{linreg}
}
\description{
\packageDescription{linreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linreg}
\packageIndices{linreg}
The linreg class provided by this package runs a linear regression by OLS on a given formula and dataset. Important fields such as residuals, predicted values and estimated coefficients can be accessed by calling their respective methods. Additionally, a summary of the model can be printed along with some basic plots. 
}
\author{
\packageAuthor{linreg}

Maintainer: \packageMaintainer{linreg}
}
\keyword{package, OLS}
\seealso{
\code{\link[<linreg>:<linreg>-package]{<linreg>}}
}