% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linreg function for Linear Regression
to perform different calculations for Linear Regression}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{is a formula}

\item{data}{is a DataFrame name
both parameters are inputs for Linear Regression different calculations
to find out Regression Coefficients and other factors}
}
\value{
the list of different objects of Linreg class
}
\description{
Linreg function for Linear Regression
to perform different calculations for Linear Regression
}
\examples{
linreg(Petal.Length~Sepal.Width+Sepal.Length,iris)
linreg(Petal.Length ~ Sepal.Length,iris)
}
