% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiU_theme.R
\name{LiU_theme}
\alias{LiU_theme}
\title{LiU Theme}
\usage{
LiU_theme()
}
\value{
A list containing the custom theme and color scale.
}
\description{
A custom theme based on the Linköping University style.
This function applies both a custom ggplot theme and a manual color scale.
}
