% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_inference.R
\name{multicoef_inference}
\alias{multicoef_inference}
\title{Hypothesis test that multiple coefficients are 0.}
\usage{
multicoef_inference(model, testeq0, alpha = 0.05)
}
\arguments{
\item{model}{returned object from fit_ols_model}

\item{testeq0}{vector of values from 1 to p, indicating which coefficients are being tested}

\item{alpha}{significance level}
}
\description{
Hypothesis test that multiple coefficients are 0.
}
