% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_estimation.R
\name{ols_fit}
\alias{ols_fit}
\title{Fit an Ordinary Least Squares Model}
\usage{
ols_fit(X, y, homo = T)
}
\arguments{
\item{X}{A matrix of predictors, if intercept included, must be on p=1.}

\item{y}{A numeric vector of responses.}

\item{homo}{Logical; TRUE assumes homoscedasticity (default), FALSE assumes heteroscedasticity.}
}
\value{
An object of class "ols" containing:
  \itemize{
    \item {X}: Input covariate matrix.
    \item {coefficients}: Estimated coefficients (\eqn{\beta}).
    \item {y_fitted}: Fitted values (\eqn{\hat{y}}).
    \item {residuals}: Residuals (\eqn{e = y - \hat{y}}).
    \item {rss}: Residual Sum of Squares.
    \item {r_squared}: Coefficient of determination (\eqn{R^2}).
    \item {adjusted_r_squared}: Adjusted \eqn{R^2}.
    \item {std_error}: Standard errors of coefficients.
    \item {hat_matrix}: Hat matrix.
    \item {leverages}: Diagonal elements of the hat matrix.
    \item {covariance_matrix}: Variance-covariance matrix of coefficients.
    \item {sigmahat_cor}: Corrected sigma estimator for homoscedasticity.
    \item {sigmahat_naive}: Naive sigma estimator.
  }
}
\description{
Fits a linear regression model using Ordinary Least Squares (OLS).
}
