% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_estimation.R
\name{ols_summary}
\alias{ols_summary}
\title{Summary for OLS Objects}
\usage{
ols_summary(model)
}
\arguments{
\item{model}{An object of fitted OLS.}
}
\value{
A summary of the OLS model fit.
}
\description{
Summary for OLS Objects
}
