% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{methods}
\name{plot,linreg-method}
\alias{plot,linreg-method}
\title{A function to plot the regression output}
\usage{
\S4method{plot}{linreg}(x)
}
\arguments{
\item{x}{a linreg object}
}
\description{
A function to plot the regression output
}
