% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{plot() method of linreg class}
\usage{
\method{plot}{linreg}(linreg.obj, standarized.residuals = FALSE)
}
\arguments{
\item{linreg.obj}{An object of linreg class}
}
\description{
Method that plot Fitted Values vs Residuals
}
\examples{
qr.obj <- QRlinreg(iris$Sepal.Length ~ iris$Sepal.Width, iris)
plot(qr.obj)

}
\references{
plot() manual - \url{https://stat.ethz.ch/R-manual/R-devel/library/base/html/plot.html}
}

