% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{plot}
\alias{plot}
\alias{plot.linreg}
\title{Plot method for linreg class}
\usage{
plot(x)

\method{plot}{linreg}(x)
}
\arguments{
\item{x}{An S3 object of class linreg.}
}
\description{
Plots the Residuals vs. Fitted & Scale-Location plots with ggplot2.
}
