% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{pred method of linreg class}
\usage{
\method{pred}{linreg}(object)
}
\arguments{
\item{object}{Linreg class object}
}
\value{
Returns the predicted values y.
}
\description{
pred method of linreg class
}
