% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{Pred method of linreg class}
\usage{
pred(x, ...)
}
\arguments{
\item{x}{Linreg class object}

\item{...}{additional parameter}
}
\value{
Returns the predicted values y.
}
\description{
Pred method of linreg class
}
