% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\name{predict}
\alias{predict}
\alias{predict.ridgereg}
\title{Predict method for ridgereg class}
\usage{
predict(x, newdata = NULL)

\method{predict}{ridgereg}(x, newdata = NULL)
}
\arguments{
\item{x}{An S3 object of class ridgereg.}

\item{newdata}{A data frame with new data (optional)}
}
\description{
Returns the preducted values.
}
