% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict method for ridgereg class}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An S3 object of class ridgereg.}

\item{newdata}{A data frame with new data (optional)}

\item{...}{Additional arguments (unused).}
}
\description{
Returns the preducted values.
}
