% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{methods}
\name{print,linreg-method}
\alias{print,linreg-method}
\title{A function to print relevant regression output}
\usage{
\S4method{print}{linreg}(x)
}
\arguments{
\item{x}{a linreg object}
}
\description{
A function to print relevant regression output
}
