% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{print() method of linreg class}
\usage{
\method{print}{linreg}(linreg.obj)
}
\arguments{
\item{linreg.obj}{An object of linreg class}
}
\description{
Method that prints out the coefficients and coefficient names, similar as done by the \code{lm} class.
}
\examples{
qr.obj <- QRlinreg(iris$Sepal.Length ~ iris$Sepal.Width, iris)
print(qr.obj)

}
\references{
print() manual - \url{https://stat.ethz.ch/R-manual/R-devel/library/base/html/print.html}
}

