% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R, R/RidgeRegression.R
\name{print}
\alias{print}
\alias{print.linreg}
\alias{print.ridgereg}
\title{Print method for linreg class}
\usage{
print(x)

\method{print}{linreg}(x)

print(x)

\method{print}{ridgereg}(x)
}
\arguments{
\item{x}{An S3 object of class ridgereg.}
}
\description{
Prints out the coefficients and coefficient names.

Prints out the coefficients and coefficient names.
}
