% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print method for ridgereg class}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{An S3 object of class ridgereg.}

\item{...}{Additional arguments (unused).}
}
\description{
Prints out the coefficients and coefficient names.
}
