% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRlinreg.R
\name{QRlinreg}
\alias{QRlinreg}
\title{QR decomposition method on Linear Regression Class}
\usage{
QRlinreg(formula, data)
}
\arguments{
\item{formula}{A formula with the independent variable and dependent variable/s}

\item{data}{A dataset}
}
\value{
An object of class \code{linreg}
}
\description{
Creates an object of class \code{linreg} with all the statistics computed for an specific \code{formula}
 and the given \code{data}
}
\examples{
QRlinreg(iris$Sepal.Length ~ iris$Sepal.Width, iris)
QRlinreg(faithful$eurption ~ faithful$waiting, faithful)

}
\references{
QR Decomposition - \url{https://en.wikipedia.org/wiki/QR_decomposition}
}

