% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Resid method of linreg class}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{Linreg class object}

\item{...}{additional parameter}
}
\value{
Returns the vector of residuals e.
}
\description{
Resid method of linreg class
}
