% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{resid}
\alias{resid}
\alias{resid.linreg}
\title{Residuals method for linreg class}
\usage{
resid(x)

\method{resid}{linreg}(x)
}
\arguments{
\item{x}{An S3 object of class linreg.}
}
\value{
A vector of residuals.
}
\description{
Residuals method for linreg class
}
