% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Linear Regression by Ridge Regression}
\description{
The ridgereg class runs a linear regression by ridge regression on a given formula and dataset.
}
\section{Fields}{

\describe{
\item{\code{formula}}{An R formula describing the desired formula for the linear regression}

\item{\code{data}}{A data.frame containing the necessary data for the linear regression}

\item{\code{lambda}}{A scalar value lambda to tune the model}

\item{\code{beta_hats}}{A matrix with the estimated coefficients}

\item{\code{y_hat}}{A matrix with the predicted values for y}

\item{\code{formula_call}}{The formula as used in the object initialization}

\item{\code{data_call}}{The data as used in the object initialization}
}}


