% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\name{ridgereg}
\alias{ridgereg}
\title{ridgereg function}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{A formula object describing the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{lambda}{A numeric value specifying the ridge regularization parameter.}
}
\value{
An S3 object of class ridgereg containing the fitted model.
}
\description{
Performs ridge regression on a given formula and dataset.
It normalizes the covariates and calculates the ridge regression coefficients.
}
