% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{significant_stars}
\alias{significant_stars}
\title{Significant Stars Function}
\usage{
significant_stars(pval)
}
\arguments{
\item{pval}{P values}
}
\value{
Returns the significant stars based on the \code{pval}
}
\description{
Significant Stars Function
}
