% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{summary() method of linreg class}
\usage{
\method{summary}{linreg}(linreg.obj)
}
\arguments{
\item{linreg.obj}{An object of linreg class}
}
\description{
Method that prints out the a summary of an object from the class linreg
}
\examples{
qr.obj <- QRlinreg(iris$Sepal.Length ~ iris$Sepal.Width, iris)
summary(qr.obj)

}
\references{
summary() manual - \url{https://stat.ethz.ch/R-manual/R-devel/library/base/html/summary.html}
}

