% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_estimation.R
\name{summary.ols}
\alias{summary.ols}
\title{Summary for OLS Objects}
\usage{
\method{summary}{ols}(model)
}
\arguments{
\item{model}{An object of fitted OLS.}
}
\value{
A summary of the OLS model fit.
}
\description{
Summary for OLS Objects
}
