% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{summary}
\usage{
summary(model)
}
\arguments{
\item{model}{a model fit through linearRegression}
}
\value{
a list containing the model formula, residuals, a data frame (entitled 'coefficients') with estimated regression coefficients, standard errors, t-statistics, and p-values for each coefficient, as well as
R-squared, adjusted R-squared, degrees of freedom, model F-statistic, p-value, and residual standard error
}
\description{
Determines various characteristics of a linearRegression model
}
\examples{
fit<-linearRegression(dist~speed, cars); summary(fit)
fit2<-linearRegression(mpg~cyl+disp+hp+wt+cyl:disp+disp:wt, dat=mtcars); summary(fit2)$coefficients
y<-rnorm(100); x<-rnorm(100); fit3<-linearRegression(y~x, subs=1:50); summary(fit3)

}
