% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{test_multiple_coefficients}
\alias{test_multiple_coefficients}
\title{Hypothesis test that multiple coefficients are 0.}
\usage{
test_multiple_coefficients(model, testeq0, alpha = 0.05)
}
\arguments{
\item{model}{returned object from fit_ols_model}

\item{testeq0}{vector of values from 1 to p, indicating which coefficients are being tested}

\item{alpha}{significance level}
}
\description{
Hypothesis test that multiple coefficients are 0.
}
