% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_inference.R
\name{test_row_constraints}
\alias{test_row_constraints}
\title{Hypothesis test for row constraints H0: R beta = r.}
\usage{
test_row_constraints(model, R, r = 0, alpha = 0.05)
}
\arguments{
\item{model}{returned object from fit_ols_model}

\item{R}{matrix of row constraints -- must have p columns, full row rank, and no inconsistent constraints}

\item{r}{value of r to use in null hypothesis H0: R beta = r}

\item{alpha}{significance level}
}
\description{
Hypothesis test for row constraints H0: R beta = r.
}
