% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_forward.R
\name{train_forward}
\alias{train_forward}
\title{Custom forward-selection}
\usage{
train_forward(y, data)
}
\description{
Custom forward-selection
}
\section{Fields}{

\describe{
\item{\code{y}}{The response variable}

\item{\code{data}}{A data.frame containing the necessary data for the linear regression, including response variable}
}}

