% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{anova}
\usage{
anova(model)
}
\arguments{
\item{model}{a model fit through linearRegression}
}
\value{
an analysis of variance table
}
\description{
Performs analysis of variance on a model already fitted through linearRegression
}
\examples{
fit<-linearRegression(dist~speed, cars); anova(fit)

}
