% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients method for linreg class}
\usage{
\method{coef}{linreg}(x)
}
\arguments{
\item{x}{An object of class linreg.}
}
\description{
Coefficients method for linreg class
}
