% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{coef}
\alias{coef}
\alias{coef.linreg}
\title{Coefficients method for linreg class}
\usage{
coef(x)

\method{coef}{linreg}(x)
}
\arguments{
\item{x}{An S3 object of class linreg.}
}
\value{
A named vector of coefficients.
}
\description{
Coefficients method for linreg class
}
