% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef}
\alias{coef}
\title{Coef method of linreg class}
\usage{
coef(x, ...)
}
\arguments{
\item{x}{Linreg class object}

\item{...}{additional parameter}
}
\value{
Returns the coefficients as a named vector.
}
\description{
Coef method of linreg class
}
