\name{linearRegression-package}
\alias{linearRegression-package}
\alias{linearRegression}
\docType{package}
\title{
\packageTitle{linearRegression}
}
\description{
\packageDescription{linearRegression}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linearRegression}
\packageIndices{linearRegression}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{linearRegression}

Maintainer: \packageMaintainer{linearRegression}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
