% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearRegression.R
\name{linearRegression}
\alias{linearRegression}
\title{linearRegression}
\usage{
linearRegression(form, dat = NULL, subs = NULL)
}
\arguments{
\item{form}{an object of class 'formula', or one that can be coerced to that class. ie. a symbolic description of the model to be fitted}

\item{dat}{an optional data frame containing the data on the model to be fit}

\item{subs}{an optional vector specifying a subset of data to be used}
}
\value{
a list of length 5 containing the coefficients of the model, residuals, fitted values, rank of the fitted model, the residual degrees of freedom, and the model frame used
}
\description{
Performs a linear regression analysis on the given formula
}
\examples{
linearRegression(dist~speed, cars)
y<-rnorm(100); x<-rnorm(100); linearRegression(y~x, subs=1:50)

}
