% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\docType{package}
\name{LinearRegression}
\alias{LinearRegression}
\alias{LinearRegression-package}
\title{LinearRegression: A package for performing a linear regression.}
\description{
LinearRegression: A package for performing a linear regression.
}
\section{LinearRegression functions}{

linreg(formula, data)
}

