% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegressionFunction.R
\name{linearregression}
\alias{linearregression}
\title{Linear Regression Function}
\usage{
linearregression(X, Y)
}
\arguments{
\item{X}{input value}

\item{Y}{input value}
}
\value{
the coefficient estimates, fitted values, and residuals
}
\description{
Conducts linear regression model and computes coefficient estimates, fitted values, and residuals.
}
\examples{
library(datasets)
data(mtcars)
Y <- mtcars$hp
X <- cbind(mtcars$wt, mtcars$drat)
linearregression(X,Y)

}
