% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression package}
\arguments{
\item{formula}{a relation between variables}

\item{datathe}{data set containing variable values}
}
\value{
a linear regression object containing relevant paramters
}
\description{
Linear Regression package
}

