% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R, R/linreg_object.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Perform linear regression}
\arguments{
\item{formula}{An formula}

\item{data}{A data.frame}
}
\value{
A linear regression in \code{RegressionValues}
}
\description{
Perform linear regression

A Reference Class to represent a linear regression
}
\section{Fields}{

\describe{
\item{\code{a}}{A numeric something}
}}


