% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg}
\value{
An object of class "linreg".
}
\description{
A Reference Class that creates a linear regression based on a given
formula with p independent variables and a given data set with n
observations. Allows users to view plots and data summaries based on this
regression.
}
\section{Fields}{

\describe{
\item{\code{regdata}}{The name of the data set being used. A character string.}

\item{\code{regformula}}{The formula being used for the regression. A character
string.}

\item{\code{coefficients}}{Coefficient estimates for each independent variable. A
matrix with one column and p+1 rows.}

\item{\code{fittedvalues}}{Fitted values of dependent variable for each observation.
A matrix with one column and n rows.}

\item{\code{residuals}}{Residual values for dependent variable for each observation.
A matrix with one column and n rows.}

\item{\code{degrees}}{Degrees of Freedom. A length-one numeric vector.}

\item{\code{residvariance}}{Residual Variance. A length-one numeric vector.}

\item{\code{coeffvariance}}{Coefficient Variance for each independent variable. A
matrix with one column and p+1 rows.}

\item{\code{tvalues}}{T-values for each coefficient estimate. A matrix with one
column and p+1 rows.}

\item{\code{pvalues}}{P-values for each coefficient estimate. A matrix with one
column and p+1 rows.}

\item{\code{sumstats}}{Summary statistics for the regression (coefficient estimates,
standard errors, t-values, and p-values). A matrix with with four columns
and p+1 rows.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Shows the coefficient estimates as a named vector.}

\item{\code{initialize(formula, data)}}{Creates the object, executes the regression calculations based on formula
and data, and saves the values under the respective fields of the object.}

\item{\code{pred()}}{Shows the fitted/predicted values for each observation in the data set.}

\item{\code{print()}}{Prints the coefficient estimates for the regression.}

\item{\code{resid()}}{Shows the residual values for each observation in the data set.}

\item{\code{summary()}}{Shows the summary statistics for the regression, including the coefficient
estimates, standard errors, t-values, p-values, residual standard error, 
and degrees of freedom.}
}}

