% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression by OLS}
\description{
The linreg class runs a linear regression by OLS on a given formula and dataset. 
Important fields such as residuals, predicted values and estimated coefficients 
can be accessed by calling their respective methods. Additionally, a summary of 
the model can be printed along with some basic plots.
}
\section{Fields}{

\describe{
\item{\code{formula}}{An R formula describing the desired formula for the linear regression}

\item{\code{data}}{A data.frame containing the necessary data for the linear regression}

\item{\code{beta_hats}}{A matrix with the estimated coefficients}

\item{\code{y_hat}}{A matrix with the predicted values for y}

\item{\code{res}}{A matrix with the residuals}

\item{\code{df}}{Degrees of freedom}

\item{\code{res_var}}{Variance of the residuals}

\item{\code{var_beta_hats}}{Variance of the estimated coefficients}

\item{\code{t_beta_hats}}{A matrix of t-values for the estimated coefficients}

\item{\code{p_beta_hats}}{A matrix of p-values for the estimated coefficients (t-test)}
}}


