% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear regression
Linear regression function.}
\arguments{
\item{formula}{= Petal.Length~Sepal.Width+Sepal.Length}

\item{data}{= iris}
}
\value{
func<-linreg$new(formula,data)

coef()

print()

resid()

plot()

summary()

pred()
}
\description{
Linear regression
Linear regression function.
}

\author{
c("Nastaran Meftahi", "Sarah Walid Alsaadi")
}
