% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression S3 class- linreg}
\usage{
linreg(formula, data, ...)
}
\arguments{
\item{formula}{Equation with dependent and Independent variables.}

\item{data}{Input data.frame}

\item{...}{}
}
\value{
Returns the linreg class object of list containing 
                        1> Regression Coefficients, 
                        2> Co-variance matrix, 
                        3> residuals, 
                        4> fitted values, 
                        5> degrees of freedom, 
                        6> variance (sigma2),
                        7> Standard Error,
                        8> T value
                        9> P value
                        10> Function call
}
\description{
Linear Regression S3 class- linreg
}
\details{
This linreg class uses S3 class as object oriented programming. Uses the function lnreg as
          regression computation engine. 
          
Methods associated with linreg class:
1> print(mod_object): Prints the regression formula and data along with the coefficients.

2> Summary(mod_object): Returns a similar printout as printed for lm objects, but includes only the coefficients with 
                       their standard error, t-value and p-value as well as the estimate of residual variance
                       and the degrees of freedom in the model.

3> coef(mod_object): Returns the coefficients as a named vector.
4> resid(mod_object): Returns the vector of residuals e.

5> pred(mod_object): Returns the predicted values y.

6> plot(mod_object): Plots the two plots (1. Residuals vs Fitted, 2. Scale−Location) using ggplot2.
}
\examples{
data(iris)
mod_object <- linreg(Petal.Length~Species, data = iris)
print(mod_object)
summary(mod_objet)
plot(mod_object)

}
