% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{LiU_theme}
\alias{LiU_theme}
\title{LiU Theme}
\usage{
LiU_theme()
}
\value{
A list containing the custom theme, color scale, and font settings.
}
\description{
A custom theme based on the LiU style.
This function applies both a custom ggplot theme and a manual color scale.
}
