% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_estimation.R
\name{ols_fit}
\alias{ols_fit}
\title{Fit an Ordinary Least Squares Model}
\usage{
ols_fit(X, y, homo = T)
}
\arguments{
\item{X}{A matrix of predictors (design matrix).}

\item{y}{A numeric vector of responses.}

\item{homo}{Logical; TRUE assumes homoscedasticity (default), FALSE assumes heteroscedasticity.}
}
\value{
An object of class "ols" containing:
  \itemize{
    \item \code{X}: Input covariate matrix.
    \item \code{coefficients}: Estimated coefficients (\eqn{\beta}).
    \item \code{y_fitted}: Fitted values (\eqn{\hat{y}}).
    \item \code{residuals}: Residuals (\eqn{e = y - \hat{y}}).
    \item \code{rss}: Residual Sum of Squares (RSS).
    \item \code{r_squared}: Coefficient of determination (\eqn{R^2}).
    \item \code{adjusted_r_squared}: Adjusted \eqn{R^2}.
    \item \code{std_error}: Standard errors of coefficients.
    \item \code{hat_matrix}: Hat matrix (\eqn{H}).
    \item \code{leverages}: Diagonal elements of the hat matrix (\eqn{H}).
    \item \code{covariance_matrix}: Variance-covariance matrix of coefficients.
    \item \code{sigmahat_cor}: Corrected sigma estimator for homoscedasticity.
    \item \code{sigmahat_naive}: Naive sigma estimator.
  }
}
\description{
Fits a linear regression model using Ordinary Least Squares (OLS).
}
