% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method for linreg class}
\usage{
\method{plot}{linreg}(x)
}
\arguments{
\item{x}{An object of class linreg.}
}
\description{
Plot method for linreg class
}
