% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method of linreg class}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{Linreg class object}

\item{...}{additional parameter}
}
\value{
Plots the two plots (1. Residuals vs Fitted, 2. Scale - Location) using ggplot2.
}
\description{
Plot method of linreg class
}
