% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted values method for linreg class}
\usage{
pred.linreg(x)
}
\arguments{
\item{x}{An object of class linreg.}
}
\description{
Predicted values method for linreg class
}
