% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Pred method of linreg class}
\usage{
\method{pred}{linreg}(x)
}
\arguments{
\item{x}{Linreg class object}

\item{...}{additional parameter}
}
\value{
Returns the predicted values y.
}
\description{
Pred method of linreg class
}
