% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{pred}
\alias{pred}
\alias{pred.linreg}
\title{Predicted values method for linreg class}
\usage{
pred(x)

\method{pred}{linreg}(x)
}
\arguments{
\item{x}{An S3 object of class linreg.}
}
\value{
A vector of predicted values.
}
\description{
Predicted values method for linreg class
}
