% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{pred method of linreg class}
\usage{
pred(object)
}
\arguments{
\item{object}{Linreg class object}
}
\value{
Returns the predicted values y.
}
\description{
pred method of linreg class
}
