% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method for linreg class}
\usage{
\method{print}{linreg}(x)
}
\arguments{
\item{x}{An object of class linreg.}
}
\description{
Print method for linreg class
}
