% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method of linreg class}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{}
}
\value{
Print out the coefficients and coefficient names, similar as done by the lm class.
}
\description{
Print method of linreg class
}
