% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method of linreg class}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{Linreg class object}

\item{...}{additional parameter}
}
\value{
Prints out the regression formula and data along with the coefficients
}
\description{
Print method of linreg class
}
