% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{print}
\alias{print}
\alias{print.linreg}
\title{Print method for linreg class}
\usage{
print(x)

\method{print}{linreg}(x)
}
\arguments{
\item{x}{An S3 object of class linreg.}
}
\description{
Prints out the coefficients and coefficient names.
}
