% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Residuals method for linreg class}
\usage{
\method{resid}{linreg}(x)
}
\arguments{
\item{x}{An object of class linreg.}
}
\description{
Residuals method for linreg class
}
