% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression Package}
\arguments{
\item{formula}{a relation between variables}

\item{data}{the data set containing variable values}

\item{lambda}{the penalty coefficient}
}
\value{
a rigid regression object containing relevant paramters
}
\description{
Ridge regression shrinks the regression coefficients by imposing a penalty on their size. The New element in this function is Lambda, the  effective degrees of freedom
}
\section{Fields}{

\describe{
\item{\code{input,}}{the input variables, formula and data}

\item{\code{var_name}}{the variable names}

\item{\code{regres_coef}}{contains the estimated beta value. If the scrutinised system contains one dependent variable, the number of element in regres_coef is two, beta_zero and beta one. On the other hand, if two dependent variables are included, the number of beta elements would be three.}

\item{\code{fitted_y,}}{is the estimated value, using the regression coefficients}

\item{\code{res_value,}}{is the residuals, the difference between the sampling points and the estimated value}

\item{\code{degrees_of_freedom,}}{the degree of freedom describe the degress of freedom of the scrutinised system}

\item{\code{res_value,}}{the residual variance is the variance of the error, between the estimated value and the sampling value}

\item{\code{standard_resvec,}}{the standard residuals gives a measure of the strength of the difference between observed and expected values}

\item{\code{res_var,}}{the variance of the regression coefficients}

\item{\code{t_values,}}{the t-values for each coefficient}

\item{\code{p_value,}}{the p-values for each coefficient}

\item{\code{sq_standard_resvec}}{gives the square root of the standard residual vector value}
}}


