% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary method for linreg class}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An S3 object of class linreg.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary method for linreg class
}
