% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{summary}
\alias{summary}
\alias{summary.linreg}
\title{Summary method for linreg class}
\usage{
summary(x)

\method{summary}{linreg}(x)
}
\arguments{
\item{x}{An S3 object of class linreg.}
}
\description{
Summary method for linreg class
}
