% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary}
\alias{summary}
\title{Summary method of linreg class}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{Linreg class object}

\item{...}{additional parameter}
}
\value{
Returns a similar printout as printed for lm objects, but includes only the coefficients with 
                       their standard error, t-value and p-value as well as the estimate of residual variance
                       and the degrees of freedom in the model
}
\description{
Summary method of linreg class
}
