% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_lm.R
\name{toy_lm}
\alias{toy_lm}
\title{Fit a simple linear regression model}
\usage{
toy_lm(formula, data)
}
\arguments{
\item{formula}{A formula specifying the model (e.g., mpg ~ wt + hp + disp)}

\item{data}{A data frame containing the variables in the model}
}
\value{
A list containing coefficients, fitted values, and residuals
}
\description{
Fit a simple linear regression model
}
\examples{
# Fit a simple model using mtcars dataset
fit <- toy_lm(mpg ~ wt + hp + disp, data = mtcars)
fit

}
