% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_lm.R
\name{toy_lm}
\alias{toy_lm}
\title{Fit a Linear Regression Model (Simplified lm)}
\usage{
toy_lm(formula, data)
}
\arguments{
\item{formula}{A model formula (e.g., mpg ~ wt + hp)}

\item{data}{A data frame containing the variables in the formula}
}
\value{
A list containing:
\item{coefficients}{Coefficient table with SE, t, p}
\item{fitted}{Fitted values}
\item{residuals}{Residuals}
\item{r.squared}{R-squared}
\item{adj.r.squared}{Adjusted R-squared}
\item{mse}{Mean squared error}
\item{residual.se}{Residual standard error}
}
\description{
A simplified implementation of linear regression using the normal equations.
Returns coefficient table, fitted values, residuals.
}
\examples{
fit <- toy_lm(mpg ~ wt + hp + disp, data = mtcars)
fit$coefficients

}
