% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize airport delays}
\usage{
visualize_airport_delays()
}
\value{
A plot
}
\description{
Creates a plot that visualizes the mean (arrival) delay of flights for different airports by longitude and latitude.
}
