% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Map of Airport Delays}
\usage{
visualize_airport_delays()
}
\description{
Plot a map of mean delays at airports throughout the US.
}
